enigo

Struct Enigo

pub struct Enigo { /* private fields */ }

Implementations

impl Enigo

pub fn new(settings: &Settings) -> Result<Self, NewConError>

Create a new Enigo struct to establish the connection to simulate input with the specified settings

Errors

Have a look at the documentation of NewConError to see under which conditions an error will be returned.

pub fn delay(&self) -> u32

Get the delay per keypress

pub fn set_delay(&mut self, delay: u32)

Set the delay per keypress

pub fn held(&mut self) -> (Vec<Key>, Vec<u16>)

Returns a list of all currently pressed keys