enigo

Struct Settings

pub struct Settings {
    pub mac_delay: u32,
    pub linux_delay: u32,
    pub x11_display: Option<String>,
    pub wayland_display: Option<String>,
    pub windows_dw_extra_info: Option<usize>,
    pub event_source_user_data: Option<i64>,
    pub release_keys_when_dropped: bool,
}

Settings for creating the Enigo struct and it’s behavior

Fields
mac_delay: u32

Sleep delay on macOS

linux_delay: u32

Sleep delay on Linux X11

x11_display: Option<String>

Display name to connect to when using Linux X11

wayland_display: Option<String>

Display name to connect to when using Linux Wayland

windows_dw_extra_info: Option<usize>

Arbitrary value to be able to distinguish events created by enigo All events will be marked with this value in the dwExtraInfo field

event_source_user_data: Option<i64>

Arbitrary value to be able to distinguish events created by enigo All events will be marked with this value in the EVENT_SOURCE_USER_DATA field

release_keys_when_dropped: bool

Set this to true if you want all held keys to get released when Enigo gets dropped